+++ backend/src/main/java/site/kkokkio/domain/comment/dto/CommentReportRequestDto.java
@@ -1,11 +1,11 @@
-package site.kkokkio.domain.report.dto;
+package site.kkokkio.domain.comment.dto;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import site.kkokkio.global.enums.ReportReason;
 
 @Builder
-public record PostReportRequestDto(
+public record CommentReportRequestDto(
         @NotNull
         ReportReason reason
 ) {
