+++ backend/src/main/java/site/kkokkio/domain/comment/entity/CommentReport.java
@@ -1,9 +1,8 @@
-package site.kkokkio.domain.report.entity;
+package site.kkokkio.domain.comment.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
 import site.kkokkio.domain.member.entity.Member;
-import site.kkokkio.domain.post.entity.Post;
 import site.kkokkio.global.enums.ReportProcessingStatus;
 import site.kkokkio.global.enums.ReportReason;
 import site.kkokkio.global.util.BaseTimeEntity;
@@ -13,22 +12,22 @@ import site.kkokkio.global.util.BaseTimeEntity;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Table(name = "post_report",
+@Table(name = "comment_report",
         uniqueConstraints = {
-                @UniqueConstraint(columnNames = {"post_id", "reporter_id"})
+                @UniqueConstraint(columnNames = {"comment_id", "reporter_id"})
         })
-public class PostReport extends BaseTimeEntity {
+public class CommentReport extends BaseTimeEntity {
 
     // 신고 ID
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "post_report_id")
+    @Column(name = "comment_report_id")
     private Long id;
 
-    // 신고된 포스트와의 관계
+    // 신고된 댓글과의 관계
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "post_id", nullable = false)
-    private Post post;
+    @JoinColumn(name = "comment_id", nullable = false)
+    private Comment comment;
 
     // 신고한 사용자와의 관계
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
@@ -41,7 +40,7 @@ public class PostReport extends BaseTimeEntity {
     @Builder.Default
     private ReportReason reason = ReportReason.BAD_CONTENT;
 
-    // 관리자 처리 상태 필드
+    // 관리자 처리 상태
     @Enumerated(EnumType.STRING)
     @Column(name = "status", nullable = false)
     @Builder.Default
