+++ backend/src/main/java/site/kkokkio/domain/post/dto/PostReportRequestDto.java
@@ -1,11 +1,11 @@
-package site.kkokkio.domain.report.dto;
+package site.kkokkio.domain.post.dto;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import site.kkokkio.global.enums.ReportReason;
 
 @Builder
-public record CommentReportRequestDto(
+public record PostReportRequestDto(
         @NotNull
         ReportReason reason
 ) {
