+++ backend/src/main/java/site/kkokkio/domain/post/service/PostService.java
@@ -23,9 +23,9 @@ import site.kkokkio.domain.post.entity.PostMetricHourly;
 import site.kkokkio.domain.post.repository.PostKeywordRepository;
 import site.kkokkio.domain.post.repository.PostMetricHourlyRepository;
 import site.kkokkio.domain.post.repository.PostRepository;
-import site.kkokkio.domain.report.dto.PostReportRequestDto;
-import site.kkokkio.domain.report.entity.PostReport;
-import site.kkokkio.domain.report.repository.PostReportRepository;
+import site.kkokkio.domain.post.dto.PostReportRequestDto;
+import site.kkokkio.domain.post.entity.PostReport;
+import site.kkokkio.domain.post.repository.PostReportRepository;
 import site.kkokkio.domain.source.entity.KeywordSource;
 import site.kkokkio.domain.source.entity.PostSource;
 import site.kkokkio.domain.source.entity.Source;
