--- backend/src/main/resources/db/migration/V2__add_report_count_to_comment.sql
@@ -1,2 +1,0 @@
-ALTER TABLE comment
-ADD COLUMN report_count INT NOT NULL DEFAULT 0;
